% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%rpygeo_*\%}
\alias{\%rpygeo_*\%}
\title{Multiplication operator}
\usage{
raster_1 \%rpygeo_*\% raster_2
}
\arguments{
\item{raster_1}{raster dataset or numeric}

\item{raster_2}{raster dataset or numeric}
}
\value{
{reticulate} object
}
\description{
Multiplication operator for map algebra. Spatial Analylist extension is requiered for map algebra.
}
\examples{

\dontrun{
# Load the ArcPy module and build environment
arcpy <- arcpy_build_env(overwrite = TRUE, workspace = "C:/workspace")

# Write raster to workspace directory
writeRater(elev, "C:/workspace/elev.tif", extensions = "Spatial")

# Create raster object
ras <- arcpy$sa$Raster("elev.tif")

# Multiply raster to itself
ras \%rpygeo_+\% ras \%>\%
  rpygeo_load()
}
}
\author{
Marc Becker
}
