% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{required_extensions}
\alias{required_extensions}
\title{Check required ArcGIS extensions}
\usage{
required_extensions(expr)
}
\arguments{
\item{expr}{A vector or list of character strings with Python geoprocessing
expressions or function names.}
}
\value{
Returns a character vector with the ArcGIS extension names
  (currently e.g. "Spatial", "3d", "geostats", "network", and/or
  "datainteroperability").
}
\description{
Internal function that checks which ArcGIS extensions have to be enabled to
  evaluate a Python expression.
}
\note{
This internal function is used by \code{rpygeo_geoprocessor}.
}
\seealso{
\code{\link{rpygeo_geoprocessor}}
}
\author{
Alexander Brenning, Fabian Polakowski
}
\keyword{database}
\keyword{interface}
