% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo.R
\name{rpygeo.Hillshade.sa}
\alias{rpygeo.Hillshade.sa}
\title{#*********************************************
# Spatial Analyst tools ---------------------
#*********************************************}
\usage{
rpygeo.Hillshade.sa(in.raster, out.raster, azimuth = 315, altitude = 45,
  model.shadows = c("NO_SHADOWS", "SHADOWS"), z.factor = 1, ...)
}
\description{
# http://webhelp.esri.com/arcgisdesktop/9.2/index.cfm?TopicName=Hillshade
}
