% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo_reticulate.R
\name{rpygeo_geoprocessor}
\alias{rpygeo_geoprocessor}
\title{ArcGIS Geoprocessor Workhorse}
\usage{
rpygeo_geoprocessor(lib, fun, args = NULL, extensions = NULL,
  overwrite = FALSE, workspace = NULL, detect_required_extension = TRUE)
}
\arguments{
\item{lib}{ArcPy R module name assigned using \code{rpygeo_build_env}}

\item{fun}{Single geoprocessing function name to be evaluated by the Python
geoprocessor.}

\item{args}{Vector or list of arguments to be passed to the function listed
in \code{fun}.}

\item{extensions}{Optional character vector listing ArcGIS extension that
should be enabled. This adds to any extensions that are eventually
detected by \code{rpygeo_required_extensions}.}

\item{overwrite}{If set to `TRUE` (default) existing ArcGIS datasets can be
overwritten.}

\item{workspace}{Path of ArcGIS workspace in which to perform the
geoprocessing.}

\item{detect_required_extensions}{Logical (default: \code{TRUE}).
Determines whether \code{\link{required_extensions}} should try to find out
which ArcGIS extensions are required to evaluate the \code{function(s)}.}
}
\value{
The function returns \code{NULL} if is was successful, or otherwise
  a ArcGIS error message.
}
\description{
This function utilizes the ArcPy site-package in R via the reticulate
  connection to perform ArcPy calculation in R. It returns error messages if
  an error appears.
}
\examples{

# Build a ArcGIS environment (assined to an R object called arcpy_m)
# and set overwrite to TRUE.
\dontrun{arcpy_m <- arcpy_build_env(overwrite = TRUE)}

# Use the ArcGIS Slope alogrithm to calulate a slope from a Digital Elveation
# Model
\dontrun{rpygeo_geoprocessor(lib = a, fun = "Slope_3d",
                             args = c("dem.tif", "output_slope.tif"))}

}
\seealso{
\code{\link{rpygeo_build_env}}
}
\author{
Alexander Brenning, Fabian Polakowski
}
