% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo_reticulate.R
\name{rpygeo_load}
\alias{rpygeo_load}
\title{Load output of ArcPy functions into R session}
\usage{
rpygeo_load(data)
}
\arguments{
\item{data}{Path to the ArcPy function output file}
}
\value{
`raster` or `sf` object
}
\description{
This function loads the output of an ArcPy function into the R session. Raster files are loaded as `raster` objects and vector files as `sf` objects. Currently .tif, .img and .shp files are supported. Usually this function is used with the pipe operator, hence the data parameter is not manually set.
}
\examples{

\dontrun{
# Load packages
library(spData)
library(dplyr)

# Load the ArcPy module and build environment
env <- arcpy_build_env(overwrite = TRUE, workspace = "C:/")

# Write raster to workspace directory
writeRater(elev, "C:/elev.tif")

# Create a slope raster and load it into the R Session
env$Slope_3d(in_raster = "elev.tif", out_raster = "slope.tif") \%>\%
  rpygeo_load() -> slope
}
}
\author{
Marc Becker
}
