% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo_reticulate.R
\name{rpygeo_save}
\alias{rpygeo_save}
\title{Save temporary file to workspace}
\usage{
rpygeo_save(data, filename)
}
\arguments{
\item{data}{Path to the temporary file}

\item{filename}{Filename with extension}
}
\value{
RasterLayer
}
\description{
This function saves temporary raster files as permanent files to the workspace. The raster format is inferred from the file extension.
}
\details{
Some ArcPy functions have no parameter to specify an output file. Instead they return an object and a temporary file is saved to the workspace. This functions writes the temporary file as a permanent file to the workspace. For supported formats s. \link[raster]{writeFormats}.
}
\examples{

\dontrun{
# Load packages
library(spData)
library(dplyr)

# Load the ArcPy module and build environment
env <- arcpy_build_env(overwrite = TRUE, workspace = "C:/workspace/")

# Write raster to workspace directory
writeRater(elev, "C:/workspace/elev.tif")

# Calculate
env$sa$Aspect(in_raster = "elev.tif") \%>\%
  rpygeo_save("aspect.tif")
}

}
\author{
Marc Becker
}
