% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{set_scratch_workspace}
\alias{set_scratch_workspace}
\title{Set scratch workspace}
\usage{
set_scratch_workspace(path)
}
\arguments{
\item{path}{path to the scratch workspace}
}
\description{
Set scratch workspace for ArcPy environment
}
\author{
Fabian Polakowski and Marc Becker
}
