% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{set_workspace}
\alias{set_workspace}
\title{Set workspace}
\usage{
set_workspace(path)
}
\arguments{
\item{path}{path to the workspace containing your data}
}
\description{
set workspace for ArcPy environment
}
\author{
Fabian Polakowski
}
