% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPyGeo.R
\docType{package}
\name{RPyGeo-package}
\alias{RPyGeo}
\alias{RPyGeo-package}
\title{RPyGeo: Info}
\description{
add description here
}
\details{
details
}
\author{
\strong{Maintainer}: Alexander Brenning \email{alexander.brenning@uni-jena.de}

Authors:
\itemize{
  \item Fabian Polakowski \email{fabian.polakowski@gmail.com}
}

}
