% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo_reticulate.R
\name{rpygeo_build_env}
\alias{rpygeo_build_env}
\title{Initialize Arcpy module in R}
\usage{
rpygeo_build_env(path = NULL, overwrite = TRUE, extension = NULL,
  pro = FALSE)
}
\arguments{
\item{path}{Root path to the Python version which contains "arcpy". If left
empty, the function
looks for `python.exe` in the most likely location (C:\Python27\)}
}
\value{
Returns arcpy module in R
}
\description{
Initialises the Python "arcpy" site-package in R with the help
  of reticulate
}
\author{
Fabian Polakowski
}
