% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo_reticulate.R
\name{rpygeo_search}
\alias{rpygeo_search}
\title{Search for ArcPy functions}
\usage{
rpygeo_search(search_term = NULL, module = NULL)
}
\arguments{
\item{search_term}{Search term. Regular expressions are possible.}

\item{module}{ArcPy or ArcGIS API module created with \code{\link{rpygeo_build_env}}.}
}
\value{
List of matching ArcPy functions
}
\description{
Search for ArcPy functions with a character string or regular expression.
}
\examples{
# Get  ArcPy or ArcGIS API module
\dontrun{arc <- rpygeo_build_env(pro = TRUE)
# Search for ArcPy functions, which contain the term 3d
rpygeo_search("3d")}
}
\seealso{
\code{\link{rpygeo_build_env}}
}
\author{
Marc Becker
}
