% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpygeo_reticulate.R
\name{rpygeo_search}
\alias{rpygeo_search}
\title{Search for ArcPy functions}
\usage{
rpygeo_search(search_term = NULL, module = NULL)
}
\arguments{
\item{search_term}{Search term. Regular expressions are possible.}

\item{module}{ArcPy or ArcGIS API modules created with \code{\link{rpygeo_build_env}}.}
}
\value{
Character vector of matching ArcPy functions
}
\description{
Search for ArcPy functions with a character string or regular expression.
}
\examples{

\dontrun{
# Load the ArcPy module and build environment
arcpy <- arcpy_build_env(overwrite = TRUE, workspace = "C:/workspace/")

# Search for ArcPy functions, which contain the term 3d
rpygeo_search("3d", arcpy)
}
}
\seealso{
\code{\link{rpygeo_build_env}}
}
\author{
Marc Becker
}
